PRSD-NONE: Moves Navigator into pageObjects directory
